Update molecule scenario
molecule has not been updated quite a while now and seems to be broken.

Added a requirements.txt and created a Makefile to make it easier to
create a python3 virtual environment to be able to run molecule (updated
.gitignore while at it).

The molecule scenario was recreated like this (after moving the
molecule directory out of the way):

  ❯ make install

  ❯ molecule --version
  molecule 3.5.1 using python 3.9
      ansible:2.9.26
      delegated:3.5.1 from molecule
      docker:1.0.2 from molecule_docker requiring collections: community.docker>=1.8.0

  ❯ molecule init scenario -r ansible-restic --driver-name docker

The .yamllint file got updated too.

(Actually, the molecule was copied and adapted from the role
ansible-role-website.)

To make molecule work again, the docker image had to be changed to
pre_build_image `geerlingguy/docker-debian10-ansible`.